Calculate duration in one step, not "gaps.size" steps
Simulate several separate durations by not counting the year following a
gap. In a normal subtraction the first year is not counted as well. This
is a shortcut to avoid splitting the duration into their part,
calculating individually and then adding the durations.

If there are longer gaps than one year, duplicate years are removed.