Allow to infer verband-ids manually
While this is an internal concern, the inferred data is needed during
authorization-checking. For new and not yet validated objects, the data
is not yet present. As this need to manually infer things arises, it
should be part of the public API.

And here we are discussing the promotion of a formerly internal
lifecycle-hook to a properly public method. Part of the inaugutory
festivities was a rename to avoid having a set_*-method.  After all,
this only incidentally sets the data, the important part is the
inference, dependening on context.

Another change is a proper, consistent return-value. Before, it only
returned the id the inferred mitgliederverband, with no information
about the also - potentially - inferred regionalverband.

With all this out of the way, one can move on to repairing the
authorization-check that prompted all this.

[was: rename method and make public]