Remove remote-call to keep CRUD simpler
Fixes #86

While the remote form seemed good, it was better to keep the historic
roles- page clean and quick. Inserting basically all group-names is a
waste of time and bandwidth. The existing group-search is better and
could be tweaked to return the same groups as before.

As a bug-fixing side-effect, the form now refreshes the page with a
success messsage or shows an error.