Save fabricated roles without validation
This is necessary here because we are manually creating "forbidden"
roles for the purpose of the spec. We would like to test the behaviour
of Role::MitgliederMitglied roles, which cannot be created normally
through the UI, but are currently created by the import.rake task for
importing historic roles. The rake task also saves the roles with
validation turned off, so we need to simulate that.